requirements.txt
Django>=1.5,<1.7
opps>=0.2.9
django-celery
PyVimeo==0.2.0
